Refactored new function classes into wrims-core module. Refactored calsurrogate.jar and other wrims-core dependencies into the third-party module.
